Support using  `backgroundColor='transparent'`
When the background color is set to `transparent`, the existing paths & text are not cleared - resulting in the draw method appending paths to the canvas on each draw. This change clears existing paths and text when the `backgroundColor` is set to `transparent` to result in a clean and correct render when the background is transparent.